+++ unlock-app/src/components/content/event/Form.tsx
@@ -68,6 +68,7 @@ export const Form = ({ onSubmit }: FormProps) => {
   const { networks } = useConfig()
   const { network, account } = useAuth()
   const [isInPerson, setIsInPerson] = useState(true)
+  const [isUnlimitedCapacity, setIsUnlimitedCapacity] = useState(false)
   const [isFree, setIsFree] = useState(true)
   const [isCurrencyModalOpen, setCurrencyModalOpen] = useState(false)
   const { mutateAsync: uploadImage, isLoading: isUploading } = useImageUpload()
@@ -94,7 +95,7 @@ export const Form = ({ onSubmit }: FormProps) => {
         ticket: {
           event_start_date: today,
           event_start_time: '',
-          event_end_date: '',
+          event_end_date: today,
           event_end_time: '',
           event_timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
           event_address: '',
@@ -112,7 +113,6 @@ export const Form = ({ onSubmit }: FormProps) => {
     formState: { errors },
     watch,
   } = methods
-
   const details = useWatch({
     control,
   })
@@ -149,7 +149,6 @@ export const Form = ({ onSubmit }: FormProps) => {
   const minEndDate = dayjs(ticket?.event_start_date).format('YYYY-MM-DD')
 
   const router = useRouter()
-
   return (
     <FormProvider {...methods}>
       <div className="grid grid-cols-[50px_1fr_50px] items-center mb-4">
@@ -319,12 +318,7 @@ export const Form = ({ onSubmit }: FormProps) => {
                       }
                     />
                     <Input
-                      {...register('metadata.ticket.event_start_time', {
-                        required: {
-                          value: true,
-                          message: 'This value is required',
-                        },
-                      })}
+                      {...register('metadata.ticket.event_start_time', {})}
                       type="time"
                       label="Start time"
                       error={
@@ -359,12 +353,7 @@ export const Form = ({ onSubmit }: FormProps) => {
                       }
                     />
                     <Input
-                      {...register('metadata.ticket.event_end_time', {
-                        required: {
-                          value: true,
-                          message: 'This value is required',
-                        },
-                      })}
+                      {...register('metadata.ticket.event_end_time', {})}
                       type="time"
                       min={minEndTime}
                       label="End time"
@@ -430,7 +419,7 @@ export const Form = ({ onSubmit }: FormProps) => {
 
                     {isInPerson && (
                       <Controller
-                        name="metadata.ticket.event_timezone"
+                        name="metadata.ticket.event_address"
                         control={control}
                         render={({ field: { onChange } }) => {
                           return <GoogleMapsAutoComplete onChange={onChange} />
@@ -449,9 +438,9 @@ export const Form = ({ onSubmit }: FormProps) => {
                 These settings can also be changed, but only by sending on-chain
                 transactions.
               </p>
-              <div className="relative flex flex-col gap-4">
+              <div className="relative flex flex-col mt-4">
                 <div className="flex items-center justify-between">
-                  <label className="px-1 mb-2 text-base" htmlFor="">
+                  <label className="" htmlFor="">
                     Currency & Price:
                   </label>
                   <ToggleSwitch
@@ -465,6 +454,7 @@ export const Form = ({ onSubmit }: FormProps) => {
                     }}
                   />
                 </div>
+
                 <div className="relative">
                   <SelectCurrencyModal
                     isOpen={isCurrencyModalOpen}
@@ -502,24 +492,40 @@ export const Form = ({ onSubmit }: FormProps) => {
                 </div>
               </div>
 
+              <div className="flex items-center justify-between mt-4">
+                <label className="" htmlFor="">
+                  Capacity:
+                </label>
+                <ToggleSwitch
+                  title="Unlimited"
+                  enabled={isUnlimitedCapacity}
+                  setEnabled={setIsUnlimitedCapacity}
+                  onChange={(enabled) => {
+                    if (enabled) {
+                      setValue('lock.maxNumberOfKeys', undefined)
+                    }
+                  }}
+                />
+              </div>
               <Input
                 {...register('lock.maxNumberOfKeys', {
                   min: 0,
                   valueAsNumber: true,
                   required: {
-                    value: true,
+                    value: !isUnlimitedCapacity,
                     message: 'Capacity is required. ',
                   },
                 })}
+                disabled={isUnlimitedCapacity}
                 autoComplete="off"
                 step={1}
                 pattern="\d+"
                 type="number"
                 placeholder="Capacity"
-                label="Capacity"
                 description={
                   'This is the maximum number of tickets for your event. '
                 }
+                error={errors.lock?.maxNumberOfKeys?.message}
               />
             </div>
           </Disclosure>
@@ -527,14 +533,14 @@ export const Form = ({ onSubmit }: FormProps) => {
           <div className="flex flex-col justify-center gap-6">
             {Object.keys(errors).length > 0 && (
               <div className="px-2 text-red-600">
-                Please make sure you complete all the required fields.
+                Please make sure you complete all the required fields.{' '}
               </div>
             )}
             <Button
               disabled={Object.keys(errors).length > 0}
               className="w-full"
             >
-              Deploy your contract
+              Create your event
             </Button>
           </div>
         </div>