+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+# 0.39.2
+
+- add `createLock` accepts keyPrice as number
+- `maxNumberOfKeys` is optional when creating a lock. If not passed will use unlimited.
+
 # 0.39.1
 
 - add `getDiscountHookValues` to web3Service